Canvas/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alpaca==1.0.0
alpaca-py==0.35.1
alpaca-trade-api==3.2.0
alpaca-trade-api-fixed==0.26
anaconda==0.0.1.1
annotated-types==0.7.0
ansicon==1.89.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
asyncio-nats-client==0.11.5
attrs==24.3.0
beautifulsoup4==4.12.3
blessed==1.20.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
deprecation==2.1.0
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.55.3
frozendict==2.4.6
frozenlist==1.5.0
html5lib==1.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jinxed==1.3.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.1.0
multitasking==0.0.11
nest-asyncio==1.6.0
newsapi==0.1.1
nltk==3.9.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pandas-datareader==0.10.0
pandas_ta==0.3.14b0
parso==0.8.4
peewee==3.17.8
pexpect==4.9.0
pillow==11.1.0
pip-review==1.3.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
readchar==4.2.1
regex==2024.11.6
requests==2.32.3
revel==0.9.1
six==1.17.0
soupsieve==2.6
sseclient-py==1.8.0
stack-data==0.6.3
termcolor==2.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
u==1.0
urllib3==2.3.0
uv==0.4.17
vaderSentiment==3.3.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
yarl==1.18.3
yfinance==0.2.51
